listening-comp/requirements.txt
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-speech
google-generativeai>=0.3.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
chromadb>=0.4.22
